Muted switching matrix status check
From a recent test, an issue with the HP 4084B Switching Matrix was discovered, which had not been observed before. The SWM status byte consistently returns either a 'Syntax error' or 'No Pin Board error'. However, based on my inspection, the SWM remains operational and functions normally despite these errors. To maintain the functionality of the majority of the system, the status check of the SWM has been skipped within the main measurement procedure until the problem is verified and fixed.